+++ src/monsteriser_monster_generator/systems/dnd5e/calculations/offensive/offensive_cr.py
@@ -20,7 +20,7 @@ class OffensiveChallengeRatingResult:
 
     """
 
-    challenge_rating: int
+    challenge_rating: float
     damage: OffensiveDamageResult
